+++ zcash_client_backend/src/fees/zip317.rs
@@ -111,6 +111,7 @@ where
         &self,
         _meta_source: &Self::MetaSource,
         _account: <Self::MetaSource as InputSource>::AccountId,
+        _target_height: TargetHeight,
         _exclude: &[<Self::MetaSource as InputSource>::NoteRef],
     ) -> Result<Self::AccountMetaT, <Self::MetaSource as InputSource>::Error> {
         Ok(())
@@ -214,6 +215,7 @@ where
         &self,
         meta_source: &Self::MetaSource,
         account: <Self::MetaSource as InputSource>::AccountId,
+        target_height: TargetHeight,
         exclude: &[<Self::MetaSource as InputSource>::NoteRef],
     ) -> Result<Self::AccountMetaT, <Self::MetaSource as InputSource>::Error> {
         let note_selector = NoteFilter::ExceedsMinValue(
@@ -222,7 +224,7 @@ where
                 .unwrap_or(SplitPolicy::MIN_NOTE_VALUE),
         );
 
-        meta_source.get_account_metadata(account, &note_selector, exclude)
+        meta_source.get_account_metadata(account, &note_selector, target_height, exclude)
     }
 
     fn compute_balance<P: consensus::Parameters, NoteRefT: Clone>(